# Hub_OpenAPI_Spec.yaml
openapi: 3.0.3
info:
  title: AI Hub API
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1

paths:

  /workspaces:
    get:
      summary: List workspaces
      responses:
        '200':
          description: List of workspaces

    post:
      summary: Create workspace
      requestBody:
        required: true
      responses:
        '200':
          description: Workspace created
  
  /workspaces/{id}:
    get:
      summary: Get workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace detail

  /workspaces/{id}/files:
    get:
      summary: List files in workspace path
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Directory listing

  /workspaces/{id}/file:
    get:
      summary: Read a file in workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File content

  /threads:
    get:
      summary: List threads
      responses:
        '200':
          description: Thread list

    post:
      summary: Create thread
      responses:
        '200':
          description: Thread created

  /threads/{id}/messages:
    get:
      summary: Get thread messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages list

  /tasks:
    get:
      summary: List tasks
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Task list

    post:
      summary: Create task
      responses:
        '200':
          description: Task created

  /tasks/{id}:
    get:
      summary: Get task detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task detail
    patch:
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task updated

  /workspaces/{id}/tasks:
    get:
      summary: List tasks for workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace task list

  /artifacts:
    get:
      summary: List artifacts
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: string
        - name: threadId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Artifact list
    post:
      summary: Create artifact
      responses:
        '200':
          description: Artifact created

  /artifacts/{id}:
    get:
      summary: Get artifact detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artifact detail

  /terminal/sessions:
    get:
      summary: List terminal sessions
      responses:
        '200':
          description: Terminal sessions list
    post:
      summary: Start terminal session
      responses:
        '200':
          description: Terminal session created

  /terminal/sessions/{id}/input:
    post:
      summary: Send input to terminal session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Input accepted

  /terminal/sessions/{id}/stop:
    post:
      summary: Stop terminal session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Terminal stopped

  /terminal/sessions/{id}/output:
    get:
      summary: Get buffered terminal output
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Terminal output entries